+++ lib/src/analyzers/lint_analyzer/rules/rules_list/avoid_empty_block/visitor.dart
@@ -1,0 +1,18 @@
+part of 'avoid_empty_block_rule.dart';
+
+class _Visitor extends RecursiveAstVisitor<void> {
+  final _emptyBlocks = <Block>[];
+
+  Iterable<Block> get emptyBlocks => _emptyBlocks;
+
+  @override
+  void visitBlock(Block node) {
+    super.visitBlock(node);
+
+    if (node.statements.isEmpty &&
+        node.parent is! CatchClause &&
+        !(node.endToken.precedingComments?.lexeme.contains('TODO') ?? false)) {
+      _emptyBlocks.add(node);
+    }
+  }
+}