+++ lib/src/analyzers/lint_analyzer/rules/rules_list/avoid_magic_number/config_parser.dart
@@ -1,0 +1,11 @@
+part of 'avoid_magic_number_rule.dart';
+
+class _ConfigParser {
+  static const _allowedConfigName = 'allowed';
+
+  static const _defaultMagicNumbers = [-1, 0, 1];
+
+  static Iterable<num> parseAllowedNumbers(Map<String, Object> config) =>
+      (config[_allowedConfigName] as Iterable?)?.cast<num>() ??
+      _defaultMagicNumbers;
+}